refactor: Converter 클래스를 infrastructure로 분리
